btf: move feature tests to separate file
No other changes intended.
